rename trait. Permission -> Role
