+++ sample/app/controllers/BaseAuthConfig.scala
@@ -1,8 +1,8 @@
 package controllers
 
 import jp.t2v.lab.play2.auth.AuthConfig
-import jp.t2v.lab.play2.auth.sample.{Permission, Account}
-import jp.t2v.lab.play2.auth.sample.Permission._
+import jp.t2v.lab.play2.auth.sample.{Role, Account}
+import jp.t2v.lab.play2.auth.sample.Role._
 import play.api.mvc.RequestHeader
 import play.api.mvc.Results._
 
@@ -13,14 +13,14 @@ trait BaseAuthConfig  extends AuthConfig {
 
   type Id = Int
   type User = Account
-  type Authority = Permission
+  type Authority = Role
 
   val idTag: ClassTag[Id] = classTag[Id]
   val sessionTimeoutInSeconds = 3600
 
   def resolveUser(id: Id)(implicit ctx: ExecutionContext) = Future.successful(Account.findById(id))
   def authorizationFailed(request: RequestHeader)(implicit ctx: ExecutionContext) = Future.successful(Forbidden("no permission"))
-  def authorize(user: User, authority: Authority)(implicit ctx: ExecutionContext) = Future.successful((user.permission, authority) match {
+  def authorize(user: User, authority: Authority)(implicit ctx: ExecutionContext) = Future.successful((user.role, authority) match {
     case (Administrator, _) => true
     case (NormalUser, NormalUser) => true
     case _ => false
