+++ sample/app/controllers/basic/Messages.scala
@@ -4,7 +4,7 @@ import controllers.stack.Pjax
 import jp.t2v.lab.play2.auth.AuthElement
 import play.api.mvc.Controller
 import views.html
-import jp.t2v.lab.play2.auth.sample.Permission._
+import jp.t2v.lab.play2.auth.sample.Role._
 
 trait Messages extends Controller with Pjax with AuthElement with AuthConfigImpl {
