+++ sample/app/controllers/csrf/PreventingCsrfSample.scala
@@ -2,7 +2,7 @@ package controllers.csrf
 
 import controllers.stack.TokenValidateElement
 import jp.t2v.lab.play2.auth.AuthElement
-import jp.t2v.lab.play2.auth.sample.Permission._
+import jp.t2v.lab.play2.auth.sample.Role._
 import play.api.data.Form
 import play.api.data.Forms._
 import play.api.mvc.Controller
