+++ sample/app/jp/t2v/lab/play2/auth/sample/Role.scala
@@ -1,0 +1,16 @@
+package jp.t2v.lab.play2.auth.sample
+
+sealed trait Role
+
+object Role {
+
+  case object Administrator extends Role
+  case object NormalUser extends Role
+
+  def valueOf(value: String): Role = value match {
+    case "Administrator" => Administrator
+    case "NormalUser"    => NormalUser
+    case _ => throw new IllegalArgumentException()
+  }
+
+}
