+++ sample/conf/db/migration/default/V1__create_tables.sql
@@ -3,5 +3,5 @@ CREATE TABLE account (
     email      varchar NOT NULL UNIQUE,
     password   varchar NOT NULL,
     name       varchar NOT NULL,
-    permission varchar NOT NULL
+    role       varchar NOT NULL
 );
